requirements.txt
astroid==2.3.3
certifi==2021.10.8
colorama==0.4.1
Django==2.2.5
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
future==0.18.2
importlib-metadata==4.11.3
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.8.0
Markdown==3.3.6
mccabe==0.6.1
mime==0.1.0
numpy==1.22.3
openpyxl==3.0.9
pandas==1.4.1
pylint==2.4.3
python-dateutil==2.8.2
pytz==2021.3
six==1.13.0
sqlparse==0.4.2
wincertstore==0.2
wrapt==1.11.2
zipp==3.7.0
